Forward port coolo's AMD64 compile fixes from BRANCH.
svn path=/trunk/kdenetwork/kopete/; revision=308817

+++ protocols/yahoo/libyahoo2/yahoo_fn.h
@@ -27,7 +27,7 @@
 struct yahoo_fn
 {
 	int type; 
-	int arg1, arg2;
+	long arg1, arg2;
 };
 
 int yahoo_xfrm( int table, int depth, int seed );
